move abis from js/json to contracts/abi
