+++ js/src/contracts/abi/gavcoin.json
