+++ js/src/contracts/abi/tokenreg.json
