+++ js/src/contracts/registry.js
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with Parity.  If not, see <http://www.gnu.org/licenses/>.
 
-import * as abis from '../json';
+import * as abis from './abi';
 
 export default class Registry {
   constructor (api) {
